Draft Milestone 4 version 4
I attempted milestone 4.  I think the goal is to pick the main frequencies out of the power spectrum and somehow use those to choose non-linear functions to use in a linear combination to fit the raw data.  I attempted this for two stars.  So far, it has not worked very well, but please feel free to improve on my work!!